Fix chart display, qty column type, healthcheck normalization
- Slack journal: send equity chart as attachment with image_url instead
  of raw URL text (renders as embedded image)
- db.py: change qty column from INTEGER to REAL for crypto fractional shares
- healthcheck.py: fix broken symbol normalization for position sync
- Cleaned up stuck FSLR and SHIB orphan positions

+++ db.py
@@ -44,7 +44,7 @@ def init_db(db_path: str | Path = DB_PATH):
                 timestamp TEXT NOT NULL,
                 symbol TEXT NOT NULL,
                 side TEXT NOT NULL,
-                qty INTEGER NOT NULL,
+                qty REAL NOT NULL,
                 entry_price REAL NOT NULL,
                 exit_price REAL,
                 stop_price REAL,
